requirements.txt
asyncio
aiohttp
throttler
bs4
sqlmodel
sqlalchemy
python-dotenv